Updated paynest system

+++ paynest/src/main/java/com/paynest/domain/Product.java
@@ -1,16 +1,19 @@
 package com.paynest.domain;
 
+import java.math.BigDecimal;
+
 //Product class
 public class Product {
     private int id;
     private String name;
-    private double price;
+    private BigDecimal price;
+
 
     // Constructor
     public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 
     // Getters
@@ -20,7 +23,7 @@ public class Product {
     public String getName() {
         return name;
     }
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
